Changes to Avizo angles for consistent typing and core XTek routines to properly set up geometry. Reader/Filter/Reconstruction modules need further testing but no longer seem to crash.

+++ trunk/Avizo/src/CCPiReconstruction/CCPiXtekAvizoPrepareFilter.h
@@ -35,7 +35,7 @@ class CCPIRECONSTRUCTION_API CCPiXtekAvizoPrepareFilter : public HxCompModule
 	float sourceToObject;
 	float sourceToDetector;
 	float maskRadius;
-	double* angles;
+	float* angles;
 	int numberOfProjections;
 	HxUniformScalarField3* createOutput(HxUniformScalarField3 *field);
 	void setParameters(HxUniformScalarField3* );
